Remove the old MasterClient210 and Protocol201
We only (pretend?) to support one version back, so we can remove a lot of the old code but keep the protocol versioning infrastructure.
